fix(automation): dedupe bridge utilities and align docs
Extract shared automation helpers, move generic traversal logic to a dedicated module, and update docs/tests to match runtime resolution behavior.

Made-with: Cursor
