+++ src/automation/createAutomationBridge.test.ts
@@ -3,20 +3,7 @@ import type { ElementNode } from '../core/index.js';
 import { createAutomationBridge } from './createAutomationBridge.js';
 import { setElementType } from './elementType.js';
 import { setFocusHostMetadata } from './focusHostMetadata.js';
-
-const createNode = (overrides: Record<string, unknown> = {}): ElementNode => {
-  const node: Partial<ElementNode> & Record<string, unknown> = {
-    id: undefined,
-    x: 0,
-    y: 0,
-    w: 0,
-    h: 0,
-    children: [],
-    parent: undefined,
-    ...overrides,
-  };
-  return node as ElementNode;
-};
+import { createNode } from './testing.js';
 
 describe('createAutomationBridge', () => {
   beforeEach(() => {
