+++ src/automation/stringUtils.ts
@@ -1,0 +1,5 @@
+export const getNonEmptyString = (value: unknown): string | undefined => {
+  if (typeof value !== 'string') return undefined;
+  const normalized = value.trim();
+  return normalized.length > 0 ? normalized : undefined;
+};
